12/14 Mandatory mission2

+++ src/main/java/com/ll/midium/domain/post/post/controller/PostController.java
@@ -5,12 +5,18 @@ import com.ll.midium.domain.post.post.entity.Post;
 import com.ll.midium.domain.post.post.service.PostService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.security.Principal;
+import java.util.List;
 
 @RequiredArgsConstructor
 @Controller
@@ -25,20 +31,23 @@ public class PostController {
     }
 
     @GetMapping("/post/write")
+    @PreAuthorize("isAuthenticated()")
     public String postWrite(PostCreateForm postCreateForm) {
         return "domain/post/post/write_form";
     }
 
     @PostMapping("/post/write")
-    public String postWrite(@Valid PostCreateForm postCreateForm, BindingResult bindingResult) {
+    @PreAuthorize("isAuthenticated()")
+    public String postWrite(@Valid PostCreateForm postCreateForm, BindingResult bindingResult, Principal principal) {
 
-        System.out.println(postCreateForm);
+        String name = principal.getName();
+
 
         if (bindingResult.hasErrors()) {
-            return "/post/write";
+            return "domain/post/post/write_form";
         }
-        postService.createPost(postCreateForm);
-        return "redirect:/";
+        postService.createPost(postCreateForm, name);
+        return "redirect:/post/list";
     }
 
     @GetMapping("/post/{id}")
@@ -50,8 +59,12 @@ public class PostController {
     }
 
     @GetMapping("/post/{id}/modify")
-    public String modifyPost(@PathVariable("id") Long id, Model model){
+    @PreAuthorize("isAuthenticated()")
+    public String modifyPost(@PathVariable("id") Long id, Model model, Principal principal){
         Post findPost = postService.findById(id);
+        if(!findPost.getAuthor().equals(principal.getName())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이없습니다.");
+        }
         model.addAttribute("post", findPost);
         return "domain/post/post/modify_form";
     }
@@ -62,12 +75,32 @@ public class PostController {
             return "/post/"+id+"/modify";
         }
         postService.modifyPost(id,postCreateForm);
-        return "redirect:/";
+        return "redirect:/post/list";
     }
 
     @GetMapping("/post/{id}/delete")
-    public String deletePost(@PathVariable("id") Long id){
+    @PreAuthorize("isAuthenticated()")
+    public String deletePost(@PathVariable("id") Long id, Principal principal){
+        Post findPost = postService.findById(id);
+        if(!findPost.getAuthor().equals(principal.getName())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이없습니다.");
+        }
         postService.delete(id);
-        return "redirect:/";
+        return "redirect:/post/list";
+    }
+
+    @GetMapping("/b/{username}")
+    public String userPost(Model model, @PathVariable("username") String username) {
+        List< Post > findAuthorPosts = postService.findByAuthor(username);
+        model.addAttribute("allPost", findAuthorPosts);
+        return "domain/post/post/user_post_page";
+    }
+
+    @GetMapping("/post/myList")
+    public String myPost(Principal principal, Model model) {
+        String name = principal.getName();
+        List< Post > myPostList = postService.findByAuthor(principal.getName());
+        model.addAttribute("allPost", myPostList);
+        return "domain/post/post/user_post_page";
     }
 }
